+++ sdk/resourcemanager/hybridcompute/armhybridcompute/models.go
@@ -10,6 +10,24 @@ package armhybridcompute
 
 import "time"
 
+// AccessRule - Access rule.
+type AccessRule struct {
+	// READ-ONLY; Name of the access rule.
+	Name *string
+
+	// READ-ONLY; Access rule properties
+	Properties *AccessRuleProperties
+}
+
+// AccessRuleProperties - Properties of an access rule.
+type AccessRuleProperties struct {
+	// READ-ONLY; Address prefixes that are allowed access.
+	AddressPrefixes []*string
+
+	// READ-ONLY; Direction of the access rule.
+	Direction *AccessRuleDirection
+}
+
 // AgentConfiguration - Configurable properties that the user can set locally via the azcmagent config command, or remotely
 // via ARM.
 type AgentConfiguration struct {
@@ -154,8 +172,9 @@ type ErrorDetail struct {
 
 // EsuKey - ESU key
 type EsuKey struct {
-	// The current status of the license profile key.
-	LicenseStatus *string
+	// The current status of the license profile key. Represented by the same integer value that is presented on the machine itself
+	// when querying the license key status.
+	LicenseStatus *int32
 
 	// SKU number.
 	SKU *string
@@ -172,7 +191,7 @@ type ExtensionValue struct {
 	// The single extension based on search criteria
 	Properties *ExtensionValueProperties
 
-	// READ-ONLY; Fully qualified resource ID for the resource. Ex - /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}
+	// READ-ONLY; Fully qualified resource ID for the resource. E.g. "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}"
 	ID *string
 
 	// READ-ONLY; The name of the resource
@@ -286,6 +305,9 @@ type LicenseDetails struct {
 	// Describes the license core type (pCore or vCore).
 	Type *LicenseCoreType
 
+	// A list of volume license details.
+	VolumeLicenseDetails []*VolumeLicenseDetails
+
 	// READ-ONLY; Describes the number of assigned licenses.
 	AssignedLicenses *int32
 
@@ -304,6 +326,9 @@ type LicenseProfileArmProductProfileProperties struct {
 	// Indicates the subscription status of the product.
 	SubscriptionStatus *LicenseProfileSubscriptionStatus
 
+	// READ-ONLY; The timestamp in UTC when the billing ends.
+	BillingEndDate *time.Time
+
 	// READ-ONLY; The timestamp in UTC when the billing starts.
 	BillingStartDate *time.Time
 
@@ -312,6 +337,9 @@ type LicenseProfileArmProductProfileProperties struct {
 
 	// READ-ONLY; The timestamp in UTC when the user enrolls the feature.
 	EnrollmentDate *time.Time
+
+	// READ-ONLY; The errors that were encountered during the feature enrollment or disenrollment.
+	Error *ErrorDetail
 }
 
 // LicenseProfileMachineInstanceView - License Profile Instance View in Machine Properties.
@@ -374,6 +402,15 @@ type LicenseProperties struct {
 
 	// READ-ONLY; The provisioning state, which only appears in the response.
 	ProvisioningState *ProvisioningState
+}
+
+// LicensesListResult - The List license operation response.
+type LicensesListResult struct {
+	// REQUIRED; The list of licenses.
+	Value []*License
+
+	// The URI to fetch the next page of Machines. Call ListNext() with this URI to fetch the next page of license profile.
+	NextLink *string
 }
 
 // LinuxParameters - Input for InstallPatches on a Linux VM, as directly received by the API
@@ -912,6 +949,12 @@ type MachineRunCommandScriptSource struct {
 	ScriptURIManagedIdentity *RunCommandManagedIdentity
 }
 
+// MachineRunCommandUpdate - Describes a Machine Extension Update.
+type MachineRunCommandUpdate struct {
+	// Resource tags
+	Tags map[string]*string
+}
+
 // MachineRunCommandsListResult - Describes the Run Commands List Result.
 type MachineRunCommandsListResult struct {
 	// The uri to fetch the next page of run commands. Call ListNext() with this to fetch the next page of run commands.
@@ -967,6 +1010,80 @@ type NetworkInterface struct {
 type NetworkProfile struct {
 	// The list of network interfaces.
 	NetworkInterfaces []*NetworkInterface
+}
+
+// NetworkSecurityPerimeter - Properties that define a Network Security Perimeter resource.
+type NetworkSecurityPerimeter struct {
+	// READ-ONLY; Azure resource Id
+	ID *string
+
+	// READ-ONLY; Regional location of the perimeter
+	Location *string
+
+	// READ-ONLY; Guid of the Network Security Perimeter
+	PerimeterGUID *string
+}
+
+// NetworkSecurityPerimeterConfiguration - Properties that define a Network Security Perimeter resource.
+type NetworkSecurityPerimeterConfiguration struct {
+	// Properties that define a Network Security Perimeter resource.
+	Properties *NetworkSecurityPerimeterConfigurationProperties
+
+	// READ-ONLY; Azure resource Id
+	ID *string
+
+	// READ-ONLY; Azure resource name
+	Name *string
+
+	// READ-ONLY; Azure resource type
+	Type *string
+}
+
+// NetworkSecurityPerimeterConfigurationListResult - A list of network security perimeter configurations.
+type NetworkSecurityPerimeterConfigurationListResult struct {
+	// READ-ONLY; Link to retrieve next page of results.
+	NextLink *string
+
+	// READ-ONLY; Array of results.
+	Value []*NetworkSecurityPerimeterConfiguration
+}
+
+// NetworkSecurityPerimeterConfigurationProperties - Properties that define a Network Security Perimeter resource.
+type NetworkSecurityPerimeterConfigurationProperties struct {
+	// The Network Security Perimeter associated with this configuration.
+	NetworkSecurityPerimeter *NetworkSecurityPerimeter
+
+	// Network Security Perimeter profile
+	Profile *NetworkSecurityPerimeterProfile
+
+	// The Resource Association.
+	ResourceAssociation *ResourceAssociation
+
+	// READ-ONLY; Provisioning issues.
+	ProvisioningIssues []*ProvisioningIssue
+
+	// READ-ONLY; Current state of this NetworkSecurityPerimeter: whether or not is has been provisioned within the resource group
+	// it is defined. Users cannot change this value but are able to read from it. Values will
+	// include Provisioning ,Succeeded, Canceled and Failed.
+	ProvisioningState *string
+}
+
+// NetworkSecurityPerimeterProfile - Network Security Perimeter profile
+type NetworkSecurityPerimeterProfile struct {
+	// READ-ONLY; Collection of access rules for the profile
+	AccessRules []*AccessRule
+
+	// READ-ONLY; Access rules version number
+	AccessRulesVersion *int32
+
+	// READ-ONLY; Diagnostic settings version number
+	DiagnosticSettingsVersion *int32
+
+	// READ-ONLY; Collection of enabled log categories for the profile
+	EnabledLogCategories []*string
+
+	// READ-ONLY; Name of the resource
+	Name *string
 }
 
 // OSProfile - Specifies the operating system settings for the hybrid machine.
@@ -1034,8 +1151,24 @@ type PatchSettings struct {
 	// Specifies the assessment mode.
 	AssessmentMode *AssessmentModeTypes
 
+	// Captures the hotpatch capability enrollment intent of the customers, which enables customers to patch their Windows machines
+	// without requiring a reboot.
+	EnableHotpatching *bool
+
 	// Specifies the patch mode.
 	PatchMode *PatchModeTypes
+
+	// READ-ONLY; Status of the hotpatch capability enrollment or disenrollment.
+	Status *PatchSettingsStatus
+}
+
+// PatchSettingsStatus - Status of the hotpatch capability enrollment or disenrollment.
+type PatchSettingsStatus struct {
+	// Indicates the current status of the hotpatch being enabled or disabled.
+	HotpatchEnablementStatus *HotpatchEnablementStatus
+
+	// READ-ONLY; The errors that were encountered during the hotpatch capability enrollment or disenrollment.
+	Error *ErrorDetail
 }
 
 // PrivateEndpointConnection - A private endpoint connection
@@ -1222,6 +1355,9 @@ type ProductFeature struct {
 	// Indicates the current status of the product features.
 	SubscriptionStatus *LicenseProfileSubscriptionStatus
 
+	// READ-ONLY; The timestamp in UTC when the billing ends.
+	BillingEndDate *time.Time
+
 	// READ-ONLY; The timestamp in UTC when the billing starts.
 	BillingStartDate *time.Time
 
@@ -1230,6 +1366,45 @@ type ProductFeature struct {
 
 	// READ-ONLY; The timestamp in UTC when the user enrolls the feature.
 	EnrollmentDate *time.Time
+
+	// READ-ONLY; The errors that were encountered during the feature enrollment or disenrollment.
+	Error *ErrorDetail
+}
+
+// ProvisioningIssue - Details on issues that occurred during provisioning.
+type ProvisioningIssue struct {
+	// READ-ONLY; Name of the provisioning issue.
+	Name *string
+
+	// READ-ONLY; Provisioning issue properties
+	Properties *ProvisioningIssueProperties
+}
+
+// ProvisioningIssueProperties - Properties of a provisioning issue.
+type ProvisioningIssueProperties struct {
+	// READ-ONLY; Description of the provisioning issue.
+	Description *string
+
+	// READ-ONLY; Issue type
+	IssueType *ProvisioningIssueType
+
+	// READ-ONLY; Severity of the provisioning issue.
+	Severity *ProvisioningIssueSeverity
+
+	// READ-ONLY; Access rules that can be added to the perimeter to remediate the issue
+	SuggestedAccessRules []*AccessRule
+
+	// READ-ONLY; ARM Ids of the resources that can be associated to the same perimeter to remediate the issue
+	SuggestedResourceIDs []*string
+}
+
+// ResourceAssociation - Properties that define a Resource Association.
+type ResourceAssociation struct {
+	// READ-ONLY; The access mode
+	AccessMode *AccessMode
+
+	// READ-ONLY; Name of the Resource Association
+	Name *string
 }
 
 // RunCommandInputParameter - Describes the properties of a run command parameter.
@@ -1306,6 +1481,14 @@ type SystemData struct {
 type TagsResource struct {
 	// Resource tags
 	Tags map[string]*string
+}
+
+type VolumeLicenseDetails struct {
+	// The invoice id for the volume license.
+	InvoiceID *string
+
+	// Describes the program year the volume license is for.
+	ProgramYear *ProgramYear
 }
 
 // WindowsParameters - Input for InstallPatches on a Windows VM, as directly received by the API